new指令过程分析

+++ src/main/java/vm/opcode/iload_1.java
@@ -16,7 +16,7 @@ public class iload_1 extends OpcodeSupport{
 
     @Override
     public Object operate(StackFrame frame){
-        frame.getOperands().push(frame.getLocals()[1]);
+        frame.getOperands().push((Integer)frame.getLocals()[1]);
         return null;
     }
 }
